ipq40xx: backport SFP debugfs support
Upstream has added a simple debugfs for SFP
ports in 5.12 and it is really usefull for seeing
that is the SFP module state without having to
recompile the kernel and add prints.

So, lets backport it to IPQ40xx for start altough
it should probably go to generic backports anyway.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>