ipq40xx: qca8k-ipq40xx: fix busy wait function
Current busy wait function is not working properly.
The condition for breaking out is incorrect, the function is checking for
the busy bit to become 1 which is incorrect per the datasheet.
You are supposed to poll for the bit to return to  0 after your operation.

So fix that and also move the sleep and timeout time into a define and
into the header file.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>