ipq40xx: qca8k-ipq40xx: drop QM code
Upon discussions with Ansuel Smith and his experience with the QCA8337
switch it seems that this code should be unnecessary.
Regular QCA8337 has it in SSDK as well but it should be triggered only
in extreme cases.

So, since its really messy code that pokes the PHY-s directly (which you
are not supposed to do) and very unlikely to ever be triggered drop it.
If its ever needed then it be re-implemented by triggering on the Phylink
provided adjust_link callback instead of this mess.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>