ipq40xx: qca8k-ipq4019: integrate basic PSGMII calibration
It looks like the basic calibration required for PSGMII to work is just
calibrate the VCO PLL-s in the SoC PSGMII PHY and the QCA807x PSGMII PHY.

The rest of "calibration" is actually fine tuning the PHY-s which from my
experience is needed like 1 out of 10 boots.

So, for starters port the basic stuff that is required and individual PHY
testing will be next.

This allows getting rid of the mac-mode DT property and potentially the
whole delayed DSA stuff as now the clock and reset management can be moved
to the ethernet driver and DSA will just deffer until ethernet is up.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>