kernel: package MHI bus support
ath11k requires MHI bus support for the PCI card support, so package
it as kernel module so it can be added as a dependency to the PCI ath11k
driver module.

Currently, only 5.10 kernel is supported as there is no support for it in
5.4 and backporting would take significant effort.

Backported are minimal changes required for MHI to work under 5.10 with
ath11k.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>