kernel: package QRTR support
QRTR is Qualcomm IPC router protocol and ath11k requires it for both
AHB and PCI support, so package it as a kernel module so it can be
added as a dependency to the ath11k module.

Like with MHI, only kernel 5.10 is supported due to no support for it
in kernel 5.4 and backporting would require significant effort.

Backported are minimal changes required for QRTR to work under 5.10 with
ath11k.

TUN and MHI variants are package while SMD one that will be required for
AHB ath11k is not yet as it requires RPMSG and SoC specific RPMSG drivers.
This will be added when AHB ath11k gets added.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>