mac80211: add PCI ath11k support
ath11k is the upstream driver for Qualcomm 802.11ax radios, both for the
internal AHB (Not covered by this patch) and PCI based cards.

It differs a bit from ath10k as it requires stuff like QRTR and MHI bus
support so those had to be packaged before this patch.

Since PCI variant is not SoC specific and cards are widely available start
by adding support for it for now.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>