ipq40xx: qca8k-ipq4019: get rid of the DSA delayed hack
Now that IPQESS is taking care of ESS clock and reset we can finally get
rid of the delayed work for DSA switch registration to avoid HW lockup
due to reset being called multiple times.

We can rely that IPQESS will enable the clock and reset the ESS as
dsa_register_switch() will check if the set upstream ethernet device is
avaialable and if not simply deffer probing and try later thus preventing
the switch to be registered as long as the ethernet device hasnt finished
probing.

Since platform_driver_probe() cannot be used with deferred probing convert
the driver to use the standard module_platform_driver and while this is
being done also add the remove method so that if built as a module the
DSA switch gets unregistered properly.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>